+++ packages/miscope/src/miscope/analysis/analyzers/__init__.py
@@ -42,6 +42,7 @@ from miscope.analysis.analyzers.parameter_trajectory_pca import ParameterTraject
 from miscope.analysis.analyzers.registry import AnalyzerRegistry
 from miscope.analysis.analyzers.repr_geometry import RepresentationalGeometryAnalyzer
 from miscope.analysis.analyzers.transient_frequency import TransientFrequencyAnalyzer
+from miscope.analysis.analyzers.weight_spectra import WeightSpectraAnalyzer
 
 
 __all__ = [
@@ -74,4 +75,5 @@ __all__ = [
     "TransientFrequencyAnalyzer",
     "NeuronGrouping",
     "ParameterDMD",
+    "WeightSpectraAnalyzer",
 ]